+++ .prettierrc.cjs
@@ -1,0 +1,9 @@
+const config = {
+  plugins: ['prettier-plugin-tailwindcss'],
+  printWidth: 100,
+  semi: false,
+  singleQuote: true,
+  arrowParens: 'avoid',
+}
+
+module.exports = config