+++ documentation/helpers/StoryLabel/index.tsx
@@ -8,8 +8,8 @@ interface StoryLabelProps {
 export const StoryLabel = ({ children, className }: PropsWithChildren<StoryLabelProps>) => (
   <p
     className={cx(
-      'px-md text-caption bg-info-container text-on-info-container mb-md last:mb-none flex w-fit rounded-sm font-bold leading-6',
-      className
+      'mb-md flex w-fit rounded-sm bg-info-container px-md text-caption font-bold leading-6 text-on-info-container last:mb-none',
+      className,
     )}
   >
     {children}