+++ packages/components/dialog/src/DialogBody.tsx
@@ -13,13 +13,13 @@ export const Body = forwardRef(
       className={cx(
         className,
         ['px-xl', 'py-md', 'flex-grow', 'overflow-y-auto'],
-        ['outline-none', 'focus-visible:ring-2', 'focus-visible:ring-outline-high']
+        ['outline-none', 'focus-visible:ring-2', 'focus-visible:ring-outline-high'],
       )}
       {...rest}
     >
       {children}
     </div>
-  )
+  ),
 )
 
 Body.displayName = 'Dialog.Body'