+++ packages/components/drawer/src/DrawerContent.styles.tsx
@@ -55,7 +55,7 @@ export const drawerContentStyles = cva(
       side: 'right',
       size: 'md',
     },
-  }
+  },
 )
 
 export type DrawerContentStylesProps = VariantProps<typeof drawerContentStyles>