+++ packages/components/alert-dialog/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.10](https://github.com/adevinta/spark/compare/@spark-ui/alert-dialog@0.2.9...@spark-ui/alert-dialog@0.2.10) (2023-08-23)
+
+**Note:** Version bump only for package @spark-ui/alert-dialog
+
+## [0.2.9](https://github.com/adevinta/spark/compare/@spark-ui/alert-dialog@0.2.8...@spark-ui/alert-dialog@0.2.9) (2023-08-23)
+
+**Note:** Version bump only for package @spark-ui/alert-dialog
+
+## [0.2.8](https://github.com/adevinta/spark/compare/@spark-ui/alert-dialog@0.2.7...@spark-ui/alert-dialog@0.2.8) (2023-08-22)
+
+**Note:** Version bump only for package @spark-ui/alert-dialog
+
+## [0.2.7](https://github.com/adevinta/spark/compare/@spark-ui/alert-dialog@0.2.6...@spark-ui/alert-dialog@0.2.7) (2023-08-22)
+
+**Note:** Version bump only for package @spark-ui/alert-dialog
+
 ## [0.2.6](https://github.com/adevinta/spark/compare/@spark-ui/alert-dialog@0.2.5...@spark-ui/alert-dialog@0.2.6) (2023-08-16)
 
 **Note:** Version bump only for package @spark-ui/alert-dialog