+++ packages/components/dialog/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.12](https://github.com/adevinta/spark/compare/@spark-ui/dialog@1.4.11...@spark-ui/dialog@1.4.12) (2023-08-23)
+
+**Note:** Version bump only for package @spark-ui/dialog
+
+## [1.4.11](https://github.com/adevinta/spark/compare/@spark-ui/dialog@1.4.10...@spark-ui/dialog@1.4.11) (2023-08-23)
+
+**Note:** Version bump only for package @spark-ui/dialog
+
+## [1.4.10](https://github.com/adevinta/spark/compare/@spark-ui/dialog@1.4.9...@spark-ui/dialog@1.4.10) (2023-08-22)
+
+**Note:** Version bump only for package @spark-ui/dialog
+
 ## [1.4.9](https://github.com/adevinta/spark/compare/@spark-ui/dialog@1.4.8...@spark-ui/dialog@1.4.9) (2023-08-16)
 
 **Note:** Version bump only for package @spark-ui/dialog