+++ packages/components/icons/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.20.1](https://github.com/adevinta/spark/compare/@spark-ui/icons@1.20.0...@spark-ui/icons@1.20.1) (2023-08-23)
+
+**Note:** Version bump only for package @spark-ui/icons
+
 # [1.20.0](https://github.com/adevinta/spark/compare/@spark-ui/icons@1.19.5...@spark-ui/icons@1.20.0) (2023-08-16)
 
 ### Features