+++ packages/components/spinner/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.2](https://github.com/adevinta/spark/compare/@spark-ui/spinner@2.1.1...@spark-ui/spinner@2.1.2) (2023-08-23)
+
+**Note:** Version bump only for package @spark-ui/spinner
+
+## [2.1.1](https://github.com/adevinta/spark/compare/@spark-ui/spinner@2.1.0...@spark-ui/spinner@2.1.1) (2023-08-23)
+
+**Note:** Version bump only for package @spark-ui/spinner
+
 # [2.1.0](https://github.com/adevinta/spark/compare/@spark-ui/spinner@2.0.5...@spark-ui/spinner@2.1.0) (2023-08-16)
 
 ### Features