+++ packages/hooks/use-combined-state/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.3](https://github.com/adevinta/spark/compare/@spark-ui/use-combined-state@0.5.2...@spark-ui/use-combined-state@0.5.3) (2023-08-23)
+
+**Note:** Version bump only for package @spark-ui/use-combined-state
+
 ## [0.5.2](https://github.com/adevinta/spark/compare/@spark-ui/use-combined-state@0.5.1...@spark-ui/use-combined-state@0.5.2) (2023-08-16)
 
 **Note:** Version bump only for package @spark-ui/use-combined-state